requirements.txt
asgiref==3.7.2
boto3==1.28.4
botocore==1.31.4
Django==4.2.3
jmespath==1.0.1
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16